+++ controller/user_controller.js
@@ -1,6 +1,26 @@
 const mongoose = require('mongoose')
 
 module.exports = {
-    register: (req, res) => {},
-    login: (req, res) => {},
+    register: (req, res) => {
+
+        //validation user_validate
+
+        //password match validation
+
+        //generate hash
+
+        //verify duplicate user
+
+        //create user account
+    },
+    login: (req, res) => {
+
+        //validation email & pw
+
+        //verify user email exists
+
+        //verify correct password
+
+        //JWT expiry
+    },
 }
