+++ middlewares/user_validate.js
@@ -1,4 +1,5 @@
 const joi = require('joi')
+const joiPhone = joi.extend(require('joi-phone-number'))
 
 module.exports = {
 
@@ -10,7 +11,9 @@ module.exports = {
     registerValidate: joi.object({
         name: joi.string().min(3).max(30).required(),
         email: joi.string().email().required(),
+        mobile: joiPhone.string().phoneNumber().required(),
         password: joi.string().min(8).required(),
-        confirmPassword: joi.string().required()
+        confirmPassword: joi.string().required(),
+        //checkbox validation? user agreement
     })
 }
