+++ models/choices.js
@@ -1,0 +1,21 @@
+const GenderMale = 'm'
+const GenderFemale = 'f'
+const EnumUserGender = [GenderMale, GenderFemale]
+
+//4 main categories
+const FitnessCategory = "Fitness";
+const AcademicCategory = "Academia";
+const ArtsCategory = "Arts & crafts";
+const CookCategory = "Cooking & Baking";
+const EnumMainCategories = [
+	FitnessCategory,
+	AcademicCategory,
+	ArtsCategory,
+	CookCategory,
+];
+
+
+module.exports = {
+    EnumUserGender: EnumUserGender,
+    EnumMainCategories: EnumMainCategories,
+}
