+++ package.json
@@ -24,6 +24,7 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "joi": "^17.4.0",
+    "joi-phone-number": "^5.0.1",
     "mongodb": "^3.6.10",
     "mongoose": "^5.13.2",
     "nodemon": "^2.0.9"
